apps/pantry/requirements.txt
# FastAPI and server
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
motor==3.6.0
pymongo==4.9.1

# Authentication
PyJWT==2.10.1
python-dotenv==1.2.1

# CORS and middleware
starlette==0.50.0

# Data validation
pydantic==2.12.5
pydantic_core==2.41.5

# QR Code generation
qrcode[pil]==8.0
Pillow==11.1.0

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-cov==6.0.0

